requirements.txt
# Core dependencies
unsloth>=2024.12
transformers>=4.46.0
datasets>=2.21.0
peft>=0.13.0
bitsandbytes>=0.44.0
accelerate>=0.34.0
trl>=0.12.0

# Validation & utils
pydantic>=2.9.0
tqdm>=4.66.0

# Data generation (Gemini free tier)
google-generativeai>=0.8.0

# Jupyter/Colab
ipywidgets>=8.1.0
